+++ iaso/api/workflows/changes.py
@@ -1,0 +1,5 @@
+from iaso.api.common import ModelViewSet, HasPermission
+
+
+class WorkflowChangeViewSet(ModelViewSet):
+    pass
