+++ iaso/urls.py
@@ -69,8 +69,9 @@ from .api.tasks.create.import_gpkg import ImportGPKGViewSet
 from .dhis2.authentication import dhis2_callback  # type: ignore
 from hat.api.token_authentication import token_auth
 
-from .api.workflows import WorkflowVersionViewSet
-from .api.mobile.workflows import MobileWorkflowViewSet
+from .api.workflows.versions import WorkflowVersionViewSet
+from .api.workflows.followups import WorkflowFollowupViewSet
+from .api.workflows.mobile import MobileWorkflowViewSet
 
 URL = Union[URLPattern, URLResolver]
 URLList = List[URL]
@@ -131,6 +132,7 @@ router.register(r"mobile/storage/logs", StorageLogViewSet, basename="storagelogs
 router.register(r"mobile/storage/blacklisted", StorageBlacklistedViewSet, basename="storageblacklisted")
 
 router.register(r"workflowversions", WorkflowVersionViewSet, basename="workflowversions")
+router.register(r"workflowfollowups", WorkflowFollowupViewSet, basename="workflowfollowups")
 router.register(r"mobile/workflows", MobileWorkflowViewSet, basename="mobileworkflows")
 
 router.registry.extend(plugins_router.registry)
